Huntflow/requirements.txt
pandas~=1.3.5
requests~=2.27.1
PyYAML~=6.0